refactor(mcp): introduce ToolRegistration, fold adapters/mcp into asap.mcp.auth, prune ASYNC_POLLING (S3)
Sprint S3 Task 4.0 — MCP Auth Bridge cleanup:

- Replace the untyped positional 5-tuple MCPServer._tools with a frozen
  ToolRegistration dataclass (name/handler/schema/metadata/capabilities);
  ProtectedMCPServer.from_server + _handle_tools_call + _validate_tools_at_startup
  now use attribute access. Capability metadata stays per-tool-instance
  (guards against the S2 #240 URN-agnostic resolve-cache regression;
  pinned by tests/adapters/mcp/test_per_urn_isolation.py).
- Add a typed JwtVerifyResult.capabilities accessor (list[str]; [] when
  absent/non-list) so the MCP bridge stops re-deriving the JWT claim with
  isinstance.
- Fold adapters/mcp/* into asap.mcp.auth/* (the adapter boundary was a
  layering inversion); adapters/mcp/* becomes a thin deprecation shim
  re-exporting protect_server/MCPAuthConfig/ProtectedMCPServer/
  resolve_jwt_extractor (identity preserved: asap.adapters.mcp.X is
  asap.mcp.auth.X). errors.py -> mcp.protocol cross-package import is
  now intra-package. M-3 hygiene: hide_unauthorized_tools no-op warning
  refined.
- Prune the dead OpenAPIExecutionKind.ASYNC_POLLING variant (grep-confirmed
  no live consumer; factory branches on STREAMING only); 202+Location now
  classifies as SYNC.

Behavior-preserving; asap-compliance + example importers repointed to
asap.mcp.auth (shim keeps legacy path working).
