+++ asap-compliance/tests/test_mcp_auth.py
@@ -8,7 +8,7 @@ from typing import Any
 
 import pytest
 
-from asap.adapters.mcp.errors import (
+from asap.mcp.auth.errors import (
     AUTH_REQUIRED,
     CAPABILITY_DENIED,
     CONSTRAINT_VIOLATION,
