+++ docs/adapters/openapi.md
@@ -36,7 +36,7 @@ OpenAPI document (URL or local .json)
 
 - **One async entrypoint**: `create_from_openapi` builds an `OpenAPIAdapterBundle` with a ready-to-use `Manifest`, `HandlerRegistry`, and metadata (`require_webauthn_for`, `upstream_base_url`).
 - **Upstream base URL**: Taken from `servers[0].url` when it is absolute (`https://…`). If the spec uses a **relative** server URL (e.g. `/api/v3`), the adapter resolves it with `urllib.parse.urljoin` against the document URL used to load the spec. Loading from a **local file only** with a relative server requires `upstream_base_url=`.
-- **Execution modes (OA-010)**: The mapper classifies each operation as `sync`, `streaming` (response advertises `text/event-stream`), or `async_polling` (`202` + `Location`). The manifest’s `Capability.streaming` flag is set if any operation is streaming; the current **`task.request` path** still performs a single HTTP round-trip and returns JSON (or wrapped non-object JSON). Treat streaming/async polling as **metadata for future wiring**, not full ASAP streaming handlers yet.
+- **Execution modes (OA-010)**: The mapper classifies each operation as `sync` or `streaming` (response advertises `text/event-stream`). The manifest’s `Capability.streaming` flag is set if any operation is streaming; the current **`task.request` path** still performs a single HTTP round-trip and returns JSON (or wrapped non-object JSON). Treat streaming as **metadata for future wiring**, not a full ASAP streaming handler yet. (The former `async_polling` variant for `202` + `Location` was pruned in v2.5.1 Sprint S3 as dead metadata — no production path consumed it; such operations now classify as `sync`.)
 - **Identity & capability grants**: `create_app` still provisions identity and capability HTTP routes. Use [`FreshSessionConfig`](../security/self-authorization-prevention.md) and `bundle.require_webauthn_for` (from `approval_strength`) for approval/WebAuthn policy; registering `CapabilityDefinition` rows in `app.state.capability_registry` is separate from OpenAPI mapping.
 
 ## Quick usage
