+++ tests/adapters/mcp/__init__.py
@@ -1,1 +1,7 @@
-"""Tests for asap.adapters.mcp."""
+"""Tests for the MCP Auth Bridge (now :mod:`asap.mcp.auth`).
+
+Sprint S3 Wave C Task 4.2 folded ``asap.adapters.mcp`` into
+:mod:`asap.mcp.auth`; these tests exercise the canonical ``asap.mcp.auth``
+implementation. The ``asap.adapters.mcp`` deprecation shim is verified
+separately by the import-resolution check.
+"""
